refactor: extract PostTranscriptionResultHandlers.swift
Extracts 2 @MainActor result-handler classes
(FinishedRecordingPostTranscriptionResultHandler, RetryPostTranscriptionResultHandler)
from PostTranscriptionPipelineController.swift into own file.

Byte-identical move; SHA verified against origin/main.

Reduces PostTranscriptionPipelineController.swift from 274 to 136 lines.

Tests: 667 (unchanged).
